Fix: desktop menu hover state mouseover ux
* The desktop menu expands with hover.
This inevitably causes issues with menus expanding and collapsing as
the cursor moves over child menus. And it's very likely that
the user won't be able to reach the lower items in the menus because of
this.

This commit places a transparent div of 200px of height at the end of
depth 1 of child menus to make sure that the lower vicinity of the
menu still triggers the hover state. This allows the menu to be
more accessible, even though this is a clearly hacky method.