Fix: Menu open/close issues on mobile
* Mobile menu buttons didn't allow switching between menus in
succession. What happened was that the first touch  closed the open
menu, followed by the second touch openeing the second menu. This issue
was fixed.

* LinkNav supports onSelect prop to drill the this to the
LinkPreloader. This is done to allow the menu items to be able to
close the menus when they are selected.

* MobileMenuContainer props have been renamed to give some clarity for
their purpose. The change has been reflected in app.router.