Fix: isLoading state on history go back
* isLoading state was not being triggered while the browser go back
feature was used. This meant that the user had no indication that the
page was reloading a previously visited page in case the cache ran out.

This error has been fixed with this commit. However there could be
cases where the state could remain triggered for a while if the user
is hitting go back multiple times in succession.