Update: footer placement
* It became evident that the footer placement was going to cause
problems with the css styling that is required for the wp content.
This meant that a minor rewrite of the pipeline that leads to the
footer being rendered.

* css vars that relate to footer sizing and placement have been created

* a new div under the Footer.layout ensures that the mobile header
is displayed with white background when the page is scrolled all the way
to the bottom

* New props added to Standard.layout for controlling the footer
position as well as the vertical margins of the page.

* Padding and margins for Standard.layout have been rewritten to ensure
that the footer is displayed correctly and
the page margins are allowed to be hidden to allow page-wise control
over the margins.

* Footer in App.router have been moved inside routes to ensure that it
doesn't appear while LoaderHtmlView is being displayed. We now rely
on React to make sure that the footer is not rendered again and again
between routes. This doesn't work perfectly as the
component gets removed when its parent is removed.

* Add FOOTER.md for explaining the variables and constraints related
to the footer.