Fix: Footer artwork clashing with GraceTop in main
* The artwork was clashing with GraceTop.view. This commit introduces
some calc right side gap to make the look of the footer at least a little bit
more appealing on smaller devices.

* Apple devices are not infering the size information unless the
min-width property is also updated. I'm not going to do that even
though I know what to do. Let this commit be a silent and pretty
ineffectual protest for the ridiculous lag that apple creates in
browser tech in order to pump its store revenue. yea!