Feat: add injection.service
* This service will be responsible for injecting js to enhance wp
rendered pages for use with react.

It can do a lot of things of course but one of the main purposes of
this service is to ensure that links works within the sphere of
react-router-dom.

It will also handle contact form actions and some wordpress related ux
issues that cannot be altered without bloating wp with plugins.

* Add ref to Singular.view for dom access from outside the component.

* Add injection service in Singular.view. This is where the actual
injection action will take place.

* Update and standardize previously added interceptLinks module to
work with the injection function.