Feat: add function for getting scrollbar instance
* Previously the scroll.util was sifting through a scrollbar record
object to get the instace for the scrollbar it wants to work with. This
has been replaced by a method in Scrollbar.view that hands the
scrollbar instance without ever exposing the record object.

This is done to make the Scrollbar.view the only module that is
concerned with the scrollbar instances.