Feat: add basic desktop share component
* ShareMobile has been renamed as ShareList as the component has been
repurposed as a mere list holder, now it has no afinitiy to screen size.

* A new component that acts as a container for ShareList has been added,
called ShareDesktopContainer. As the name suggests, this container
houses the share content, similar to what MobileMenuContainer does for
small screen menus.

This module is in no way final. Actually, even the placement of the
component on the screen may wildly change depending on client feedback.
Hence, components that visually clash with the current placement of
the component have not been altered.

* App.router has been updated to reflect the naming changes done for
ShareList.view.